Refactor: dedupe shared 9x3/6x9 calibration; drop redundant K cast
- Extract the 8-frames-on-120 sensor and frame arrays into file-local
  macros so 9x3 and 6x9 share a single source of truth. Prevents the two
  formats from drifting apart on a future calibration change (which is
  exactly how the bug being fixed got there in the first place).
- Drop static_cast<float>(K) in calculateEV100 — K is now const float
  since the previous commit, so the cast is dead code.

+++ Firmware/src/lightmeter_logic.cpp
@@ -110,7 +110,7 @@ float calculateEV100(float lux)
     return NAN;
   }
 
-  return log2f((lux * 100.0f) / static_cast<float>(K));
+  return log2f((lux * 100.0f) / K);
 }
 
 void formatShutterSpeed(float lux, float aperture, int iso, char *buffer, size_t bufferSize)
